add grouped telli v2 node
