+++ nodes/Telli/Telli.node.ts
@@ -1,68 +1,22 @@
 import {
-	GenericValue,
-	IDataObject,
 	IExecuteFunctions,
-	IHttpRequestMethods,
-	ILoadOptionsFunctions,
 	INodeExecutionData,
-	INodePropertyOptions,
 	INodeType,
 	INodeTypeDescription,
 	NodeOperationError,
+	IDataObject,
 } from 'n8n-workflow';
 
-const API_ROOT_URL = 'https://api.telli.com';
-const BASE_API_URL = `${API_ROOT_URL}/v1`;
+const BASE_API_URL = 'https://api.telli.com/v1';
 
 export class Telli implements INodeType {
-
-	methods = {
-		loadOptions: {
-			async getAgents(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const response = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-					method: 'GET',
-					url: `${API_ROOT_URL}/v2/agents`,
-					headers: {
-						'Content-Type': 'application/json',
-					},
-					qs: {
-						limit: 100,
-					},
-				});
-
-				if (!response || typeof response !== 'object' || !('data' in response)) {
-					return [];
-				}
-
-				const agents = (response as IDataObject).data;
-
-				if (!Array.isArray(agents)) {
-					return [];
-				}
-
-				return agents
-					.filter((agent): agent is IDataObject => !!agent && typeof agent === 'object' && !Array.isArray(agent))
-					.map((agent) => {
-						const id = typeof agent.id === 'string' ? agent.id : '';
-						const title = typeof agent.title === 'string' && agent.title ? agent.title : 'Untitled Agent';
-
-						return {
-							name: id ? `${title} (${id})` : title,
-							value: id,
-						};
-					})
-					.filter((agent) => !!agent.value);
-			},
-		},
-	};
-
 	description: INodeTypeDescription = {
 		displayName: 'telli',
 		name: 'telli',
 		icon: 'file:telli.svg',
 		group: ['transform'],
 		version: 1,
-		subtitle: '={{$parameter["resource"] + ": " + $parameter["operation"]}}',
+		subtitle: '={{$parameter["operation"]}}',
 		description: 'Interact with the telli API',
 		defaults: {
 			name: 'telli',
@@ -77,121 +31,49 @@ export class Telli implements INodeType {
 		],
 		properties: [
 			{
-				displayName: 'Resource',
-				name: 'resource',
-				type: 'options',
-				noDataExpression: true,
-				options: [
-					{ name: 'Agent', value: 'agent' },
-					{ name: 'Call', value: 'call' },
-					{ name: 'Contact', value: 'contact' },
-					{ name: 'Contact Property', value: 'contactProperty' },
-					{ name: 'Deprecated', value: 'deprecated' },
-					{ name: 'Phone Number', value: 'phoneNumber' },
-					{ name: 'Utility', value: 'utility' },
-				],
-				default: 'contact',
-			},
-			{
 				displayName: 'Operation',
 				name: 'operation',
 				type: 'options',
 				noDataExpression: true,
-				displayOptions: { show: { resource: ['agent'] } },
 				options: [
-					{ name: 'Get Agent', value: 'get-agent-v2', action: 'Get an agent', description: 'Get an agent by ID' },
-					{ name: 'List Agents', value: 'list-agents-v2', action: 'List agents' },
-				],
-				default: 'list-agents-v2',
-			},
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: { show: { resource: ['call'] } },
-				options: [
-					{ name: 'Get Call', value: 'get-call-v1', action: 'Get a call', description: 'Get call details by call ID' },
-					{ name: 'Initiate Call', value: 'initiate-call-v1', action: 'Initiate a call', description: 'Immediately initiate an outbound call' },
-					{ name: 'Remove From Auto Dialer', value: 'remove-from-auto-dialer', action: 'Remove a contact from auto dialer', description: 'Remove a contact from the auto dialer queue' },
-					{ name: 'Remove Multiple Contacts From Auto Dialer', value: 'remove-from-auto-dialer-batch-v1', action: 'Remove multiple contacts from auto dialer' },
-					{ name: 'Schedule Call', value: 'schedule-telli-call', action: 'Schedule a call with telli', description: 'Schedule a call with telli' },
-					{ name: 'Schedule Multiple Calls', value: 'schedule-calls-batch-v1', action: 'Schedule multiple calls', description: 'Schedule multiple calls in one request' },
-				],
-				default: 'schedule-telli-call',
-			},
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: { show: { resource: ['contact'] } },
-				options: [
-					{ name: 'Create Contact', value: 'create-contact-v2', action: 'Create a contact', description: 'Create a contact using the v2 API' },
-					{ name: 'Delete Contact', value: 'delete-contact-v2', action: 'Delete a contact', description: 'Delete a contact using the v2 API' },
-					{ name: 'Get Contact', value: 'get-contact-v2', action: 'Get a contact', description: 'Get a contact by ID using the v2 API' },
-					{ name: 'Get Contact By External ID', value: 'get-contact-by-external-id-v2', action: 'Get a contact by external ID', description: 'Get a contact by external ID using the v2 API' },
-					{ name: 'List Contacts', value: 'list-contacts-v2', action: 'List contacts', description: 'List contacts using the v2 API' },
-					{ name: 'Update Contact', value: 'update-contact-v2', action: 'Update a contact', description: 'Update a contact using the v2 API' },
-				],
-				default: 'list-contacts-v2',
-			},
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: { show: { resource: ['contactProperty'] } },
-				options: [
-					{ name: 'Create Contact Property', value: 'create-contact-property-v2', action: 'Create a contact property', description: 'Create a contact property' },
-					{ name: 'Get Contact Property', value: 'get-contact-property-v2', action: 'Get a contact property', description: 'Get a contact property by key' },
-					{ name: 'List Contact Properties', value: 'list-contact-properties-v2', action: 'List contact properties' },
-					{ name: 'Update Contact Property', value: 'update-contact-property-v2', action: 'Update a contact property', description: 'Update a contact property' },
-				],
-				default: 'list-contact-properties-v2',
-			},
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: { show: { resource: ['deprecated'] } },
-				options: [
-					{ name: 'Add Contact', value: 'add-telli-contact', action: 'Add a new contact to telli', description: 'Add a new contact to telli' },
-					{ name: 'Add Contacts (Batch)', value: 'add-telli-contacts-batch', action: 'Add contacts in batch', description: 'Add multiple contacts using the v1 batch endpoint' },
-					{ name: 'Delete Contact', value: 'delete-telli-contact', action: 'Delete a contact from telli', description: 'Permanently delete a contact from telli' },
-					{ name: 'Get Contact', value: 'get-telli-contact', action: 'Get a contact', description: 'Get a contact by ID using the v1 endpoint' },
-					{ name: 'Get Contact By External ID', value: 'get-contact-by-external-id', action: 'Get contact by external ID', description: 'Retrieve detailed contact information using external contact ID' },
-					{ name: 'Get Contacts (Batch)', value: 'get-telli-contacts-batch', action: 'Get contacts in batch', description: 'Get multiple contacts using the v1 batch endpoint' },
-					{ name: 'Update Contact', value: 'update-telli-contact', action: 'Update a contact in telli', description: 'Update an existing contact in telli' },
-					{ name: 'Update Contacts (Batch)', value: 'update-telli-contacts-batch', action: 'Update contacts in batch', description: 'Update multiple contacts using the v1 batch endpoint' },
+					{
+						name: 'Add Contact',
+						value: 'add-telli-contact',
+						description: 'Add a new contact to telli',
+						action: 'Add a new contact to telli',
+					},
+					{
+						name: 'Delete Contact',
+						value: 'delete-telli-contact',
+						description: 'Permanently delete a contact from telli',
+						action: 'Delete a contact from telli',
+					},
+					{
+						name: 'Get Contact By External ID',
+						value: 'get-contact-by-external-id',
+						description: 'Retrieve detailed contact information using external contact ID',
+						action: 'Get contact by external ID',
+					},
+					{
+						name: 'Remove From Auto Dialer',
+						value: 'remove-from-auto-dialer',
+						description: 'Remove a contact from the auto dialer queue',
+						action: 'Remove a contact from auto dialer',
+					},
+					{
+						name: 'Schedule Call',
+						value: 'schedule-telli-call',
+						description: 'Schedule a call with telli',
+						action: 'Schedule a call with telli',
+					},
+					{
+						name: 'Update Contact',
+						value: 'update-telli-contact',
+						description: 'Update an existing contact in telli',
+						action: 'Update a contact in telli',
+					},
 				],
 				default: 'add-telli-contact',
-			},
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: { show: { resource: ['phoneNumber'] } },
-				options: [
-					{ name: 'Delete Phone Number', value: 'delete-phone-number-v1', action: 'Delete a phone number', description: 'Delete a phone number' },
-					{ name: 'Import Phone Number', value: 'import-phone-number-v1', action: 'Import a phone number', description: 'Import a phone number via SIP trunk' },
-					{ name: 'List Phone Numbers', value: 'list-phone-numbers-v1', action: 'List phone numbers', description: 'List active phone numbers' },
-					{ name: 'Replace Phone Number', value: 'replace-phone-number-v1', action: 'Replace a phone number', description: 'Replace a phone number with a new one' },
-				],
-				default: 'list-phone-numbers-v1',
-			},
-			{
-				displayName: 'Operation',
-				name: 'operation',
-				type: 'options',
-				noDataExpression: true,
-				displayOptions: { show: { resource: ['utility'] } },
-				options: [
-					{ name: 'Make API Call', value: 'make-api-call-v1', action: 'Make an API call', description: 'Perform an arbitrary API call to telli' },
-				],
-				default: 'make-api-call-v1',
 			},
 
 			// add-contact
@@ -529,20 +411,17 @@ export class Telli implements INodeType {
 				description: 'ID of the contact to schedule call with',
 			},
 			{
-				displayName: 'Agent Name or ID',
+				displayName: 'Agent ID',
 				name: 'agentId',
-				type: 'options',
-				typeOptions: {
-					loadOptionsMethod: 'getAgents',
-				},
-				options: [],
+				type: 'string',
 				default: '',
 				displayOptions: {
 					show: {
 						operation: ['schedule-telli-call'],
 					},
 				},
-				description: 'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
+				required: true,
+				description: 'Optional agent ID to use for the call. If not set, the default agent will be used.',
 			},
 			{
 				displayName: 'Max Retry Days',
@@ -568,33 +447,6 @@ export class Telli implements INodeType {
 				},
 				description: 'Override the from number for the call',
 			},
-			{
-				displayName: 'Call Message',
-				name: 'callMessage',
-				type: 'string',
-				default: '',
-				typeOptions: {
-					rows: 3,
-				},
-				displayOptions: {
-					show: {
-						operation: ['schedule-telli-call'],
-					},
-				},
-				description: 'Optional message added to call_details for the scheduled call',
-			},
-			{
-				displayName: 'Call Questions (JSON)',
-				name: 'callQuestions',
-				type: 'json',
-				default: '[]',
-				displayOptions: {
-					show: {
-						operation: ['schedule-telli-call'],
-					},
-				},
-				description: 'Optional call_details.questions array for the scheduled call',
-			},
 
 			// get-contact-by-external-id
 			{
@@ -609,572 +461,15 @@ export class Telli implements INodeType {
 					},
 				},
 				required: true,
-
 				description: 'Your unique internal identifier for the contact',
-			},
-
-			// v2 contacts and agents list limits
-			{
-				displayName: 'Limit',
-				name: 'listLimit',
-				type: 'number',
-				default: 50,
-				typeOptions: {
-					minValue: 1,
-					maxValue: 100,
-				},
-				displayOptions: {
-					show: {
-						operation: ['list-contacts-v2', 'list-agents-v2'],
-					},
-				},
-				description: 'Maximum number of results to return',
-			},
-
-
-			{
-				displayName: 'Agent Name or ID',
-				name: 'agentResourceId',
-				type: 'options',
-				typeOptions: {
-					loadOptionsMethod: 'getAgents',
-				},
-				options: [],
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['get-agent-v2'],
-					},
-				},
-				required: true,
-				description: 'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
-			},
-			// shared UUID identifier
-			{
-				displayName: 'ID',
-				name: 'resourceId',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: [
-							'get-contact-v2',
-							'update-contact-v2',
-							'delete-contact-v2',
-							'get-agent-v2',
-							'get-call-v1',
-							'replace-phone-number-v1',
-							'delete-phone-number-v1',
-						],
-					},
-				},
-				required: true,
-				description: 'Resource identifier',
-			},
-
-			// v2 contact fields
-			{
-				displayName: 'First Name',
-				name: 'v2FirstName',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'Contact first name',
-			},
-			{
-				displayName: 'Last Name',
-				name: 'v2LastName',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'Contact last name',
-			},
-			{
-				displayName: 'Phone Number',
-				name: 'v2PhoneNumber',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'Contact phone number in E.164 format',
-			},
-			{
-				displayName: 'External ID',
-				name: 'v2ExternalId',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2', 'get-contact-by-external-id-v2'],
-					},
-				},
-				description: 'External contact identifier',
-			},
-			{
-				displayName: 'External URL',
-				name: 'v2ExternalUrl',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'Link to contact in external system',
-			},
-			{
-				displayName: 'Salutation',
-				name: 'v2Salutation',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'Formal greeting',
-			},
-			{
-				displayName: 'Timezone (IANA)',
-				name: 'v2TimezoneIana',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'IANA timezone identifier (for example Europe/Berlin)',
-			},
-			{
-				displayName: 'Email',
-				name: 'v2Email',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'Contact email address',
-			},
-			{
-				displayName: 'Properties (JSON)',
-				name: 'v2Properties',
-				type: 'json',
-				default: '{}',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-v2', 'update-contact-v2'],
-					},
-				},
-				description: 'Custom properties object',
-			},
-
-			// old v1 get contact
-			{
-				displayName: 'Contact ID',
-				name: 'deprecatedContactId',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['get-telli-contact'],
-					},
-				},
-				required: true,
-				description: 'The telli contact ID to retrieve',
-			},
-
-			// batch payloads
-			{
-				displayName: 'Contacts (JSON)',
-				name: 'contactsBatch',
-				type: 'json',
-				default: '[]',
-				displayOptions: {
-					show: {
-						operation: ['add-telli-contacts-batch', 'update-telli-contacts-batch'],
-					},
-				},
-				description: 'Array of contacts for batch operations',
-			},
-			{
-				displayName: 'Contact IDs (JSON)',
-				name: 'contactIdsBatch',
-				type: 'json',
-				default: '[]',
-				displayOptions: {
-					show: {
-						operation: ['get-telli-contacts-batch', 'remove-from-auto-dialer-batch-v1'],
-					},
-				},
-				description: 'Array of contact IDs',
-			},
-			{
-				displayName: 'Limit',
-				name: 'batchLimit',
-				type: 'number',
-				default: 10,
-				typeOptions: {
-					minValue: 1,
-				},
-				displayOptions: {
-					show: {
-						operation: ['get-telli-contacts-batch'],
-					},
-				},
-				description: 'Max number of returned contacts for batch get',
-			},
-			{
-				displayName: 'Calls (JSON)',
-				name: 'scheduleCallsBatch',
-				type: 'json',
-				default: '[]',
-				displayOptions: {
-					show: {
-						operation: ['schedule-calls-batch-v1'],
-					},
-				},
-				description: 'Array of call scheduling objects',
-			},
-
-			// call operations
-			{
-				displayName: 'Contact ID',
-				name: 'callContactId',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['initiate-call-v1'],
-					},
-				},
-				required: true,
-				description: 'Contact identifier',
-			},
-			{
-				displayName: 'Agent Name or ID',
-				name: 'callAgentId',
-				type: 'options',
-				typeOptions: {
-					loadOptionsMethod: 'getAgents',
-				},
-				options: [],
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['initiate-call-v1'],
-					},
-				},
-				description: 'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
-			},
-			{
-				displayName: 'Max Retry Days',
-				name: 'callMaxRetryDays',
-				type: 'number',
-				default: 0,
-				typeOptions: {
-					minValue: 0,
-				},
-				displayOptions: {
-					show: {
-						operation: ['initiate-call-v1'],
-					},
-				},
-				description: 'Optional max retry days',
-			},
-			{
-				displayName: 'Override From Number',
-				name: 'callOverrideFromNumber',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['initiate-call-v1'],
-					},
-				},
-				description: 'Optional phone number override',
-			},
-
-			// contact property operations
-			{
-				displayName: 'Property Key',
-				name: 'contactPropertyKey',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['get-contact-property-v2', 'update-contact-property-v2', 'create-contact-property-v2'],
-					},
-				},
-				description: 'Unique key of the contact property',
-			},
-			{
-				displayName: 'Data Type',
-				name: 'contactPropertyDataType',
-				type: 'options',
-				options: [
-					{ name: 'Boolean', value: 'boolean' },
-					{ name: 'Date', value: 'date' },
-					{ name: 'Datetime', value: 'datetime' },
-					{ name: 'Email', value: 'email' },
-					{ name: 'Multi Select', value: 'multi_select' },
-					{ name: 'Number', value: 'number' },
-					{ name: 'Phone Number', value: 'phone_number' },
-					{ name: 'Select', value: 'select' },
-					{ name: 'String', value: 'string' },
-				],
-				default: 'string',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-property-v2'],
-					},
-				},
-				required: true,
-				description: 'Property data type',
-			},
-			{
-				displayName: 'Label',
-				name: 'contactPropertyLabel',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-property-v2', 'update-contact-property-v2'],
-					},
-				},
-				description: 'Human readable label',
-			},
-			{
-				displayName: 'Description',
-				name: 'contactPropertyDescription',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-property-v2', 'update-contact-property-v2'],
-					},
-				},
-				description: 'Property description',
-			},
-			{
-				displayName: 'Default Value',
-				name: 'contactPropertyDefaultValue',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-property-v2', 'update-contact-property-v2'],
-					},
-				},
-				description: 'Default property value',
-			},
-			{
-				displayName: 'Options (JSON)',
-				name: 'contactPropertyOptions',
-				type: 'json',
-				default: '[]',
-				displayOptions: {
-					show: {
-						operation: ['create-contact-property-v2', 'update-contact-property-v2'],
-					},
-				},
-				description: 'Options array for select and multi_select data types',
-			},
-
-			// phone number operations
-			{
-				displayName: 'Phone Number',
-				name: 'importPhoneNumber',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['import-phone-number-v1'],
-					},
-				},
-				required: true,
-				description: 'Phone number to import in E.164 format',
-			},
-			{
-				displayName: 'Termination URI',
-				name: 'importTerminationUri',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['import-phone-number-v1'],
-					},
-				},
-				required: true,
-				description: 'SIP trunk termination URI',
-			},
-			{
-				displayName: 'Auth Username',
-				name: 'importAuthUsername',
-				type: 'string',
-				default: '',
-				displayOptions: {
-					show: {
-						operation: ['import-phone-number-v1'],
-					},
-				},
-				required: true,
-				description: 'SIP trunk auth username',
-			},
-			{
-				displayName: 'Auth Password',
-				name: 'importAuthPassword',
-				type: 'string',
-				default: '',
-				typeOptions: {
-					password: true,
-				},
-				displayOptions: {
-					show: {
-						operation: ['import-phone-number-v1'],
-					},
-				},
-				required: true,
-				description: 'SIP trunk auth password',
-			},
-
-			// make api call
-			{
-				displayName: 'Path',
-				name: 'apiPath',
-				type: 'string',
-				default: '/v1/get-contact/{contactId}',
-				displayOptions: {
-					show: {
-						operation: ['make-api-call-v1'],
-					},
-				},
-				required: true,
-				description: 'Relative API path starting with /v1 or /v2',
-			},
-			{
-				displayName: 'Method',
-				name: 'apiMethod',
-				type: 'options',
-				options: [
-					{ name: 'DELETE', value: 'DELETE' },
-					{ name: 'GET', value: 'GET' },
-					{ name: 'PATCH', value: 'PATCH' },
-					{ name: 'POST', value: 'POST' },
-					{ name: 'PUT', value: 'PUT' },
-				],
-				default: 'GET',
-				displayOptions: {
-					show: {
-						operation: ['make-api-call-v1'],
-					},
-				},
-				required: true,
-				description: 'HTTP method',
-			},
-			{
-				displayName: 'Headers',
-				name: 'apiHeaders',
-				type: 'fixedCollection',
-				default: {},
-				typeOptions: {
-					multipleValues: true,
-				},
-				displayOptions: {
-					show: {
-						operation: ['make-api-call-v1'],
-					},
-				},
-				options: [
-					{
-						name: 'items',
-						displayName: 'Header',
-						values: [
-							{ displayName: 'Key', name: 'key', type: 'string', default: '' },
-							{ displayName: 'Value', name: 'value', type: 'string', default: '' },
-						],
-					},
-				],
-				description: 'Additional request headers. Authorization is added automatically.',
-			},
-			{
-				displayName: 'Query Parameters',
-				name: 'apiQueryParams',
-				type: 'fixedCollection',
-				default: {},
-				typeOptions: {
-					multipleValues: true,
-				},
-				displayOptions: {
-					show: {
-						operation: ['make-api-call-v1'],
-					},
-				},
-				options: [
-					{
-						name: 'items',
-						displayName: 'Parameter',
-						values: [
-							{ displayName: 'Key', name: 'key', type: 'string', default: '' },
-							{ displayName: 'Value', name: 'value', type: 'string', default: '' },
-						],
-					},
-				],
-				description: 'Query string parameters',
-			},
-			{
-				displayName: 'Body (JSON)',
-				name: 'apiBody',
-				type: 'json',
-				default: '{}',
-				displayOptions: {
-					show: {
-						operation: ['make-api-call-v1'],
-					},
-				},
-				description: 'Optional request body',
-			},
+			}
 		],
-
 	};
 
 	async execute(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
 		const items = this.getInputData();
 		const outputData: INodeExecutionData[] = [];
 		const operation = this.getNodeParameter('operation', 0) as string;
-
-		const parseJsonInput = (value: unknown): GenericValue | GenericValue[] | undefined => {
-			if (typeof value === 'string') {
-				const trimmed = value.trim();
-				if (!trimmed) return undefined;
-				return JSON.parse(trimmed) as GenericValue | GenericValue[];
-			}
-			return value as GenericValue | GenericValue[] | undefined;
-		};
-
-		const parseKeyValueList = (collection: IDataObject): IDataObject => {
-			const result: IDataObject = {};
-			const list = (collection.items as IDataObject[]) || [];
-			for (const item of list) {
-				if (item.key) {
-					result[item.key as string] = item.value;
-				}
-			}
-			return result;
-		};
 
 		for (let i = 0; i < items.length; i++) {
 			try {
@@ -1397,380 +692,6 @@ export class Telli implements INodeType {
 							json: getContactResponse,
 						});
 						break;
-
-
-					case 'create-contact-v2': {
-						const createContactData: IDataObject = {
-							firstName: this.getNodeParameter('v2FirstName', i, '') as string,
-							lastName: this.getNodeParameter('v2LastName', i, '') as string,
-							phoneNumber: this.getNodeParameter('v2PhoneNumber', i, '') as string,
-						};
-						if (!createContactData.firstName || !createContactData.lastName || !createContactData.phoneNumber) {
-							throw new NodeOperationError(this.getNode(), 'First Name, Last Name and Phone Number are required for Create Contact', { itemIndex: i });
-						}
-						const v2ExternalId = this.getNodeParameter('v2ExternalId', i, '') as string;
-						const v2ExternalUrl = this.getNodeParameter('v2ExternalUrl', i, '') as string;
-						const v2Salutation = this.getNodeParameter('v2Salutation', i, '') as string;
-						const v2TimezoneIana = this.getNodeParameter('v2TimezoneIana', i, '') as string;
-						const v2Email = this.getNodeParameter('v2Email', i, '') as string;
-						const v2Properties = parseJsonInput(this.getNodeParameter('v2Properties', i, '{}')) as IDataObject | undefined;
-						if (v2ExternalId) createContactData.externalId = v2ExternalId;
-						if (v2ExternalUrl) createContactData.externalUrl = v2ExternalUrl;
-						if (v2Salutation) createContactData.salutation = v2Salutation;
-						if (v2TimezoneIana) createContactData.timezoneIana = v2TimezoneIana;
-						if (v2Email) createContactData.email = v2Email;
-						if (v2Properties && typeof v2Properties === 'object' && Object.keys(v2Properties).length > 0) createContactData.properties = v2Properties;
-						const createContactResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${API_ROOT_URL}/v2/contacts`,
-							headers: { 'Content-Type': 'application/json' },
-							body: createContactData,
-						});
-						outputData.push({ json: createContactResponse });
-						break;
-					}
-
-					case 'update-contact-v2': {
-						const updateV2Id = this.getNodeParameter('resourceId', i) as string;
-						const updateContactV2Data: IDataObject = {};
-						const updateV2FirstName = this.getNodeParameter('v2FirstName', i, '') as string;
-						const updateV2LastName = this.getNodeParameter('v2LastName', i, '') as string;
-						const updateV2PhoneNumber = this.getNodeParameter('v2PhoneNumber', i, '') as string;
-						const updateV2ExternalId = this.getNodeParameter('v2ExternalId', i, '') as string;
-						const updateV2ExternalUrl = this.getNodeParameter('v2ExternalUrl', i, '') as string;
-						const updateV2Salutation = this.getNodeParameter('v2Salutation', i, '') as string;
-						const updateV2TimezoneIana = this.getNodeParameter('v2TimezoneIana', i, '') as string;
-						const updateV2Email = this.getNodeParameter('v2Email', i, '') as string;
-						const updateV2Properties = parseJsonInput(this.getNodeParameter('v2Properties', i, '{}')) as IDataObject | undefined;
-						if (updateV2FirstName) updateContactV2Data.firstName = updateV2FirstName;
-						if (updateV2LastName) updateContactV2Data.lastName = updateV2LastName;
-						if (updateV2PhoneNumber) updateContactV2Data.phoneNumber = updateV2PhoneNumber;
-						if (updateV2ExternalId) updateContactV2Data.externalId = updateV2ExternalId;
-						if (updateV2ExternalUrl) updateContactV2Data.externalUrl = updateV2ExternalUrl;
-						if (updateV2Salutation) updateContactV2Data.salutation = updateV2Salutation;
-						if (updateV2TimezoneIana) updateContactV2Data.timezoneIana = updateV2TimezoneIana;
-						if (updateV2Email) updateContactV2Data.email = updateV2Email;
-						if (updateV2Properties && typeof updateV2Properties === 'object' && Object.keys(updateV2Properties).length > 0) updateContactV2Data.properties = updateV2Properties;
-						const updateContactV2Response = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'PATCH',
-							url: `${API_ROOT_URL}/v2/contacts/${updateV2Id}`,
-							headers: { 'Content-Type': 'application/json' },
-							body: updateContactV2Data,
-						});
-						outputData.push({ json: updateContactV2Response });
-						break;
-					}
-
-					case 'delete-contact-v2': {
-						const deleteContactV2Id = this.getNodeParameter('resourceId', i) as string;
-						const deleteContactV2Response = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'DELETE',
-							url: `${API_ROOT_URL}/v2/contacts/${deleteContactV2Id}`,
-							headers: { 'Content-Type': 'text/plain' },
-						});
-						outputData.push({ json: deleteContactV2Response });
-						break;
-					}
-
-					case 'list-contacts-v2': {
-						const contactsLimit = this.getNodeParameter('listLimit', i, 50) as number;
-						const listContactsResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${API_ROOT_URL}/v2/contacts`,
-							headers: { 'Content-Type': 'application/json' },
-							qs: { limit: contactsLimit },
-						});
-						outputData.push({ json: listContactsResponse });
-						break;
-					}
-
-					case 'get-contact-v2': {
-						const getContactV2Id = this.getNodeParameter('resourceId', i) as string;
-						const getContactV2Response = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${API_ROOT_URL}/v2/contacts/${getContactV2Id}`,
-							headers: { 'Content-Type': 'application/json' },
-						});
-						outputData.push({ json: getContactV2Response });
-						break;
-					}
-
-					case 'get-contact-by-external-id-v2': {
-						const externalIdV2 = this.getNodeParameter('v2ExternalId', i) as string;
-						const getContactByExternalIdV2Response = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${API_ROOT_URL}/v2/external/contacts/${externalIdV2}`,
-							headers: { 'Content-Type': 'application/json' },
-						});
-						outputData.push({ json: getContactByExternalIdV2Response });
-						break;
-					}
-
-					case 'list-agents-v2': {
-						const agentsLimit = this.getNodeParameter('listLimit', i, 50) as number;
-						const listAgentsResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${API_ROOT_URL}/v2/agents`,
-							headers: { 'Content-Type': 'application/json' },
-							qs: { limit: agentsLimit },
-						});
-						outputData.push({ json: listAgentsResponse });
-						break;
-					}
-
-					case 'get-agent-v2': {
-						const getAgentId = this.getNodeParameter('agentResourceId', i) as string;
-						const getAgentResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${API_ROOT_URL}/v2/agents/${getAgentId}`,
-							headers: { 'Content-Type': 'application/json' },
-						});
-						outputData.push({ json: getAgentResponse });
-						break;
-					}
-
-					case 'get-call-v1': {
-						const callId = this.getNodeParameter('resourceId', i) as string;
-						const getCallResponseV1 = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${BASE_API_URL}/get-call/${callId}`,
-						});
-						outputData.push({ json: getCallResponseV1 });
-						break;
-					}
-
-					case 'initiate-call-v1': {
-						const initiateCallPayload: IDataObject = {
-							contact_id: this.getNodeParameter('callContactId', i) as string,
-						};
-						const callAgentId = this.getNodeParameter('callAgentId', i, '') as string;
-						const callOverrideFromNumberV1 = this.getNodeParameter('callOverrideFromNumber', i, '') as string;
-						const callMaxRetryDaysV1 = this.getNodeParameter('callMaxRetryDays', i, 0) as number;
-						if (callAgentId) initiateCallPayload.agent_id = callAgentId;
-						if (callOverrideFromNumberV1) initiateCallPayload.override_from_number = callOverrideFromNumberV1;
-						if (callMaxRetryDaysV1 > 0) initiateCallPayload.max_retry_days = callMaxRetryDaysV1;
-						const initiateCallResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${BASE_API_URL}/initiate-call`,
-							headers: { 'Content-Type': 'application/json' },
-							body: initiateCallPayload,
-						});
-						outputData.push({ json: initiateCallResponse });
-						break;
-					}
-
-					case 'schedule-calls-batch-v1': {
-						const scheduleCalls = parseJsonInput(this.getNodeParameter('scheduleCallsBatch', i, '[]'));
-						const scheduleCallsBatchResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${BASE_API_URL}/schedule-calls-batch`,
-							headers: { 'Content-Type': 'application/json' },
-							body: { contacts: scheduleCalls },
-						});
-						outputData.push({ json: scheduleCallsBatchResponse });
-						break;
-					}
-
-					case 'list-contact-properties-v2': {
-						const listContactPropertiesResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${API_ROOT_URL}/v2/properties/contacts`,
-							headers: { 'Content-Type': 'application/json' },
-						});
-						outputData.push({ json: listContactPropertiesResponse });
-						break;
-					}
-
-					case 'get-contact-property-v2': {
-						const getPropertyKey = this.getNodeParameter('contactPropertyKey', i) as string;
-						const getContactPropertyResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${API_ROOT_URL}/v2/properties/contacts/${getPropertyKey}`,
-							headers: { 'Content-Type': 'application/json' },
-						});
-						outputData.push({ json: getContactPropertyResponse });
-						break;
-					}
-
-					case 'create-contact-property-v2': {
-						const createPropertyPayload: IDataObject = {
-							key: this.getNodeParameter('contactPropertyKey', i) as string,
-							dataType: this.getNodeParameter('contactPropertyDataType', i) as string,
-						};
-						const createPropertyLabel = this.getNodeParameter('contactPropertyLabel', i, '') as string;
-						const createPropertyDescription = this.getNodeParameter('contactPropertyDescription', i, '') as string;
-						const createPropertyDefaultValue = this.getNodeParameter('contactPropertyDefaultValue', i, '') as string;
-						const createPropertyOptions = parseJsonInput(this.getNodeParameter('contactPropertyOptions', i, '[]'));
-						if (createPropertyLabel) createPropertyPayload.label = createPropertyLabel;
-						if (createPropertyDescription) createPropertyPayload.description = createPropertyDescription;
-						if (createPropertyDefaultValue) createPropertyPayload.defaultValue = createPropertyDefaultValue;
-						if (Array.isArray(createPropertyOptions) && createPropertyOptions.length > 0) createPropertyPayload.options = createPropertyOptions;
-						const createContactPropertyResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${API_ROOT_URL}/v2/properties/contacts`,
-							headers: { 'Content-Type': 'application/json' },
-							body: createPropertyPayload,
-						});
-						outputData.push({ json: createContactPropertyResponse });
-						break;
-					}
-
-					case 'update-contact-property-v2': {
-						const updatePropertyKey = this.getNodeParameter('contactPropertyKey', i) as string;
-						const updatePropertyPayload: IDataObject = {};
-						const updatePropertyLabel = this.getNodeParameter('contactPropertyLabel', i, '') as string;
-						const updatePropertyDescription = this.getNodeParameter('contactPropertyDescription', i, '') as string;
-						const updatePropertyDefaultValue = this.getNodeParameter('contactPropertyDefaultValue', i, '') as string;
-						const updatePropertyOptions = parseJsonInput(this.getNodeParameter('contactPropertyOptions', i, '[]'));
-						if (updatePropertyLabel) updatePropertyPayload.label = updatePropertyLabel;
-						if (updatePropertyDescription) updatePropertyPayload.description = updatePropertyDescription;
-						if (updatePropertyDefaultValue) updatePropertyPayload.defaultValue = updatePropertyDefaultValue;
-						if (Array.isArray(updatePropertyOptions) && updatePropertyOptions.length > 0) updatePropertyPayload.options = updatePropertyOptions;
-						const updateContactPropertyResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'PATCH',
-							url: `${API_ROOT_URL}/v2/properties/contacts/${updatePropertyKey}`,
-							headers: { 'Content-Type': 'application/json' },
-							body: updatePropertyPayload,
-						});
-						outputData.push({ json: updateContactPropertyResponse });
-						break;
-					}
-
-					case 'list-phone-numbers-v1': {
-						const listPhoneNumbersResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${BASE_API_URL}/phone-numbers`,
-						});
-						outputData.push({ json: listPhoneNumbersResponse });
-						break;
-					}
-
-					case 'replace-phone-number-v1': {
-						const replacePhoneNumberId = this.getNodeParameter('resourceId', i) as string;
-						const replacePhoneNumberResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${BASE_API_URL}/phone-numbers/${replacePhoneNumberId}/replace`,
-							body: {},
-						});
-						outputData.push({ json: replacePhoneNumberResponse });
-						break;
-					}
-
-					case 'import-phone-number-v1': {
-						const importPhoneNumberPayload = {
-							phoneNumber: this.getNodeParameter('importPhoneNumber', i) as string,
-							terminationUri: this.getNodeParameter('importTerminationUri', i) as string,
-							authUsername: this.getNodeParameter('importAuthUsername', i) as string,
-							authPassword: this.getNodeParameter('importAuthPassword', i) as string,
-						};
-						const importPhoneNumberResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${BASE_API_URL}/phone-numbers/import`,
-							headers: { 'Content-Type': 'application/json' },
-							body: importPhoneNumberPayload,
-						});
-						outputData.push({ json: importPhoneNumberResponse });
-						break;
-					}
-
-					case 'delete-phone-number-v1': {
-						const deletePhoneNumberId = this.getNodeParameter('resourceId', i) as string;
-						const deletePhoneNumberResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'DELETE',
-							url: `${BASE_API_URL}/phone-numbers/${deletePhoneNumberId}`,
-							headers: { 'Content-Type': 'text/plain' },
-						});
-						outputData.push({ json: deletePhoneNumberResponse });
-						break;
-					}
-
-					case 'remove-from-auto-dialer-batch-v1': {
-						const removeBatchIds = parseJsonInput(this.getNodeParameter('contactIdsBatch', i, '[]'));
-						const removeBatchResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${BASE_API_URL}/remove-from-auto-dialer-batch`,
-							headers: { 'Content-Type': 'application/json' },
-							body: { contact_ids: removeBatchIds },
-						});
-						outputData.push({ json: removeBatchResponse });
-						break;
-					}
-
-					case 'make-api-call-v1': {
-						const apiPath = this.getNodeParameter('apiPath', i) as string;
-						const apiMethod = this.getNodeParameter('apiMethod', i) as IHttpRequestMethods;
-						const apiHeadersCollection = this.getNodeParameter('apiHeaders', i, { items: [] }) as IDataObject;
-						const apiQueryCollection = this.getNodeParameter('apiQueryParams', i, { items: [] }) as IDataObject;
-						const apiBody = parseJsonInput(this.getNodeParameter('apiBody', i, '{}'));
-						const apiHeaders = parseKeyValueList(apiHeadersCollection);
-						const apiQs = parseKeyValueList(apiQueryCollection);
-						const customApiResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: apiMethod,
-							url: `${API_ROOT_URL}${apiPath}`,
-							headers: apiHeaders,
-							qs: apiQs,
-							body: apiBody,
-						});
-						outputData.push({ json: customApiResponse });
-						break;
-					}
-
-					case 'add-telli-contacts-batch': {
-						const addContactsBatch = parseJsonInput(this.getNodeParameter('contactsBatch', i, '[]'));
-						const addContactsBatchResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${BASE_API_URL}/add-contacts-batch`,
-							headers: { 'Content-Type': 'application/json' },
-							body: { contacts: addContactsBatch },
-						});
-						outputData.push({ json: addContactsBatchResponse });
-						break;
-					}
-
-					case 'update-telli-contacts-batch': {
-						const updateContactsBatch = parseJsonInput(this.getNodeParameter('contactsBatch', i, '[]'));
-						const updateContactsBatchResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'PATCH',
-							url: `${BASE_API_URL}/update-contacts-batch`,
-							headers: { 'Content-Type': 'application/json' },
-							body: { contacts: updateContactsBatch },
-						});
-						outputData.push({ json: updateContactsBatchResponse });
-						break;
-					}
-
-					case 'get-telli-contact': {
-						const getContactIdV1 = this.getNodeParameter('deprecatedContactId', i) as string;
-						const getContactV1Response = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'GET',
-							url: `${BASE_API_URL}/get-contact/${getContactIdV1}`,
-						});
-						outputData.push({ json: getContactV1Response });
-						break;
-					}
-
-					case 'get-telli-contacts-batch': {
-						const getContactsBatchIds = parseJsonInput(this.getNodeParameter('contactIdsBatch', i, '[]'));
-						const getContactsBatchLimit = this.getNodeParameter('batchLimit', i, 10) as number;
-						const getContactsBatchResponse = await this.helpers.httpRequestWithAuthentication.call(this, 'telliApi', {
-							method: 'POST',
-							url: `${BASE_API_URL}/get-contacts-batch`,
-							headers: { 'Content-Type': 'application/json' },
-							body: { contact_ids: getContactsBatchIds },
-						});
-						if (Array.isArray(getContactsBatchResponse)) {
-							for (const contact of getContactsBatchResponse.slice(0, getContactsBatchLimit)) {
-								if (contact && typeof contact === 'object' && !Array.isArray(contact)) {
-									outputData.push({ json: contact as IDataObject });
-								} else {
-									outputData.push({ json: { value: contact as GenericValue } });
-								}
-							}
-						} else {
-							outputData.push({ json: getContactsBatchResponse });
-						}
-						break;
-					}
 
 					default:
 						throw new NodeOperationError(this.getNode(), `The operation "${operation}" is not supported!`);
